+++ Pipfile
@@ -19,9 +19,14 @@ SQLAlchemy = "==1.3.19"
 Werkzeug = "==1.0.1"
 WTForms = "==2.3.3"
 Flask-JWT-Extended = "==3.24.1"
-flask-migrate = "*"
 email-validator = "*"
-flask-login = "*"
+Flask-Migrate = "==2.5.3"
+Flask-Login = "==0.5.0"
+alembic = "==1.4.3"
+python-dateutil = "==2.8.1"
+python-editor = "==1.0.4"
+Mako = "==1.1.3"
+PyJWT = "==1.7.1"
 
 [dev-packages]
 psycopg2-binary = "==2.8.6"
