+++ app/models/__init__.py
@@ -1,2 +1,3 @@
 from .db import db
 from .user import User
+from .post import Post
