+++ react-app/public/favicon.ico
