+++ react-app/src/components/NavBar.js
