+++ react-app/src/components/auth/LogoutButton.js
