+++ react-app/src/components/auth/ProtectedRoute.js
